Final Project Commit #1
I was able to implement the fetch API methods for the necessary APIs for the user and post objects.  I created a database for the user object and had to create another table for my follow/unfollow functionality. I didn't realize how involved it would be because I got the functionality to work, but I was unable to figure out how to have specific users follow specific other users. I did a little research and it said I would have to use cookies to implement that which is a little more than I can accomplish right now. I still have to create a database for the post object and then I should be done.